tuflashmac/出图 mac 客户端/requirements.txt
# TEMU 智能出图系统 V8.0
# Nano Banana Pro 版本

streamlit>=1.40.0
Pillow>=10.4.0
google-genai>=1.0.0
python-dotenv>=1.0.0
